+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.2.0
+  - Upgrade baseimage: light-baseimage:0.2.1
+  - Makefile with build no cache
+
 ## 0.1.9
   - Upgrade baseimage: light-baseimage:0.1.5